+++ lib/process-services/src/lib/app-list/apps-list.component.spec.ts
@@ -20,7 +20,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { AppsProcessService } from '../services/apps-process.service';
 import { of, throwError } from 'rxjs';
-import { defaultApp, deployedApps, nonDeployedApps } from '../mock/apps-list.mock';
+import { defaultApp, deployedApps, nonDeployedApps } from '../testing/mock/apps-list.mock';
 import { AppsListComponent, APP_LIST_LAYOUT_GRID, APP_LIST_LAYOUT_LIST } from './apps-list.component';
 import { ProcessTestingModule } from '../testing/process.testing.module';
 import { HarnessLoader } from '@angular/cdk/testing';
