+++ lib/process-services/src/lib/process-list/components/process-instance-details/process-instance-details.component.spec.ts
@@ -20,8 +20,8 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { of } from 'rxjs';
 import { CommentProcessService } from '../../../process-comments/services/comment-process.service';
-import { exampleProcess, exampleProcessNoName, mockRunningProcess, processEnded } from '../../../mock';
-import { mockProcessInstanceComments } from '../../../mock/process/process-comments.mock';
+import { exampleProcess, exampleProcessNoName, mockRunningProcess, processEnded } from '../../../testing/mock';
+import { mockProcessInstanceComments } from '../../../testing/mock/process/process-comments.mock';
 import { ProcessService } from '../../services/process.service';
 import { ProcessInstanceDetailsComponent } from './process-instance-details.component';
 import { ProcessTestingModule } from '../../../testing/process.testing.module';
