+++ lib/process-services/src/lib/process-list/components/process-instance-header/process-instance-header.component.spec.ts
@@ -17,7 +17,7 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { AppConfigService } from '@alfresco/adf-core';
-import { exampleProcess } from '../../../mock';
+import { exampleProcess } from '../../../testing/mock';
 import { ProcessInstanceHeaderComponent } from './process-instance-header.component';
 import { ProcessTestingModule } from '../../../testing/process.testing.module';
